moved rdd trait as seperate sub module
